fix: port to modern macOS (Apple Silicon + Tahoe SDK)
Make `make` produce a clean, warning-free build on macOS 14+ and Apple
Silicon without requiring any environment flags. Verified against
macOS 26.4.1 / arm64 with end-to-end DHT + friend connectivity.

Build system (cfg/systems/Darwin.mk, Makefile):
* Derive Homebrew prefix from `brew --prefix` so the same build works on
  both Intel (/usr/local) and Apple Silicon (/opt/homebrew).
* Inject openal-soft's keg-only pkg-config path into PKG_CONFIG so
  freealut (which depends on openal) resolves automatically.
* Default DISABLE_X11=1 and DISABLE_DESKTOP_NOTIFY=1 on Darwin.
  libnotify is Linux-only. X11 is used by the video-call preview/display
  windows in video_device.c, but x11focus.h is `#ifndef __APPLE__`-gated
  while toxic.h references X11_Focus whenever -DX11 is set -- so enabling
  X11 on macOS does not build cleanly. Treating both as off keeps macOS
  as a working audio-only client.
* Drop redundant explicit -l flags that duplicate pkg-config output
  (silences "ld: ignoring duplicate libraries" warnings).
* Drop the unused glib include path; fix a latent
  PKG_CONFIG_PATH = $(shell export ...) bug that produced an empty
  string.
* Move -D_GNU_SOURCE into the Linux branch of the top-level Makefile
  and remove the per-file #define _GNU_SOURCE guards (the macro is a
  no-op on macOS, so it triggered -Wunused-macros for every file that
  defined it).

macOS video capture (src/osx_video.m, src/osx_video.h):
* Replace deprecated +[AVCaptureDevice devicesWithMediaType:] with
  AVCaptureDeviceDiscoverySession (BuiltInWideAngleCamera + External +
  ContinuityCamera).
* Define COREVIDEO_SILENCE_GL_DEPRECATION before any header that
  imports CoreVideo, to silence the OpenGL CVxxxGetTypeID warnings.
* Drop the spurious [device release] (the device came from
  -[NSArray objectAtIndex:] and was not retained) -- this was a real
  over-release that would have crashed on dealloc.
* Move [input release] until after we read input.ports so the code no
  longer relies on AVCaptureSession's implicit retain to stay valid.
* Annotate bgrxtoyuv420 as static (it is file-local) and remove the
  unimplemented `bgrtoyuv420` declaration from the header (typo: the
  declared symbol never existed).
* Give osx_video_release a proper (void) prototype.
* Drop a const-stripping cast on strcpy() and another on CFRetain().

OpenAL headers (src/notify.c, src/audio_call.c, src/conference.c,
src/groupchats.c):
* Use openal-soft's <AL/al.h> on all platforms instead of macOS's
  deprecated <OpenAL/al.h> framework headers -- we already link
  openal-soft, so this just makes the include side match. Silences
  ~20 deprecation warnings.

Misc (src/main.c):
* Cast tv.tv_usec to (long) so the %06ld format string works on
  macOS, where suseconds_t is int rather than long.

Docs (INSTALL.md):
* Rename "OS X Notes" -> "macOS Notes" (and fix the TOC anchor).
* Replace the dead Tox/homebrew-tox third-party tap with the
  mainline `brew install toxcore`.
* Drop the manual PKG_CONFIG_PATH export (now handled by Darwin.mk)
  and the libnotify install (Linux-only).
* State that video calls are not currently supported on macOS and
  explain why (toxic.h / x11focus.h `__APPLE__` gating mismatch),
  rather than pointing users at a build flag that does not work.

Note: this change does not attempt to make video calls functional on
macOS. The capture side is in place; the preview/display side would
need a native Cocoa/Metal window subsystem, which is a separate effort.

+++ src/main.c
@@ -167,7 +167,7 @@ static void cb_toxcore_logger(Tox *tox, Tox_Log_Level level, const char *file, u
 
     format_time_str(timestamp, sizeof(timestamp), "%F %T", &tmp);
 
-    fprintf(fp, "%c %s.%06ld %s:%u(%s) - %s\n", tox_log_level_show(level)[0], timestamp, tv.tv_usec, file,
+    fprintf(fp, "%c %s.%06ld %s:%u(%s) - %s\n", tox_log_level_show(level)[0], timestamp, (long)tv.tv_usec, file,
             line, func, message);
 
     fflush(fp);
